+++ CHANGES.md
@@ -9,6 +9,10 @@ E-Maj - Change log
   * When computing log statistics or generating SQL statements for changes
     dump or replay, raise a warning if the requested marks range is not
     contained by a single log session.
+  * Enforce the protection against E-Maj rollback operations crossing back
+    tables group structure changes. The existing protection for the first
+    rollback operation now concerns all subsequent operations crossing again 
+    any group structure change.
   * Minor code changes and optimizations.
 
 ### Bug fixes: ###